pyproject.toml
[tool.poetry]
name = "todo-django"
version = "0.1.0"
description = ""
authors = ["abcd9999971 <abcd9999971@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
djangorestframework = "^3.15.2"
flake8 = "^7.1.2"
black = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
